backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
drf-spectacular>=0.27,<0.28
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24
psycopg2-binary>=2.9
python-decouple>=3.8
dj-database-url>=2.1,<3.0
gunicorn>=21.0
whitenoise>=6.6
PyJWT[crypto]>=2.8
cryptography>=41.0
python-dateutil>=2.8
requests>=2.31
